requirements.txt
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
Pillow==10.0.1
psycopg2==2.9.7
python-dotenv==1.0.0
pytz==2023.3.post1
randominfo==2.0.2
requests==2.31.0
soupsieve==2.5
urllib3==2.0.5
